zad1

+++ src/Point.java
@@ -1,7 +1,7 @@
 public class Point {
-    public final int x,y;
+    public final double x,y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
